requirements.txt
streamlit>=1.28.0
openai>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.1.0
langchain-community>=0.1.0
faiss-cpu>=1.7.0
pypdf2>=3.0.0
python-docx>=1.1.0
markdown>=3.0